+++ pnet-data-api-java/src/main/java/pnet/data/api/company/CompanyDataSearch.java
@@ -7,11 +7,12 @@ import pnet.data.api.util.AggregateNumberPerBrand;
 import pnet.data.api.util.AggregateNumberPerContractType;
 import pnet.data.api.util.AggregateNumberPerTenant;
 import pnet.data.api.util.AggregateNumberPerType;
+import pnet.data.api.util.CompanyMergable;
 import pnet.data.api.util.Pair;
 import pnet.data.api.util.RestrictBrand;
-import pnet.data.api.util.CompanyMergable;
 import pnet.data.api.util.RestrictContractType;
 import pnet.data.api.util.RestrictCountryCode;
+import pnet.data.api.util.RestrictDatedBackUntil;
 import pnet.data.api.util.RestrictLocation;
 import pnet.data.api.util.RestrictTenant;
 import pnet.data.api.util.RestrictType;
@@ -26,7 +27,7 @@ public class CompanyDataSearch
     extends AbstractSearchWithAggregations<CompanyItemDTO, CompanyAggregationsDTO, CompanyDataSearch>
     implements RestrictTenant<CompanyDataSearch>, RestrictBrand<CompanyDataSearch>,
     RestrictCountryCode<CompanyDataSearch>, RestrictType<CompanyDataSearch>, RestrictContractType<CompanyDataSearch>,
-    RestrictLocation<CompanyDataSearch>, CompanyMergable<CompanyDataSearch>,
+    RestrictLocation<CompanyDataSearch>, RestrictDatedBackUntil<CompanyDataSearch>, CompanyMergable<CompanyDataSearch>,
     AggregateNumberPerTenant<CompanyDataSearch>, AggregateNumberPerBrand<CompanyDataSearch>,
     AggregateNumberPerType<CompanyDataSearch>, AggregateNumberPerContractType<CompanyDataSearch>
 {