+++ pnet-data-api-java/src/main/java/pnet/data/api/person/PersonDataFind.java
@@ -12,13 +12,13 @@ import pnet.data.api.util.RestrictAdvisorAssignmentCompanyId;
 import pnet.data.api.util.RestrictAdvisorAssignmentCompanyNumber;
 import pnet.data.api.util.RestrictAdvisorAssignmentDivision;
 import pnet.data.api.util.RestrictAdvisorAssignmentType;
-import pnet.data.api.util.RestrictRole;
 import pnet.data.api.util.RestrictBrand;
 import pnet.data.api.util.RestrictCompany;
 import pnet.data.api.util.RestrictCompanyId;
 import pnet.data.api.util.RestrictCompanyNumber;
 import pnet.data.api.util.RestrictControllingArea;
 import pnet.data.api.util.RestrictCostCenter;
+import pnet.data.api.util.RestrictDatedBackUntil;
 import pnet.data.api.util.RestrictEmail;
 import pnet.data.api.util.RestrictFunction;
 import pnet.data.api.util.RestrictGuid;
@@ -28,6 +28,7 @@ import pnet.data.api.util.RestrictNumbersType;
 import pnet.data.api.util.RestrictPersonnelDepartment;
 import pnet.data.api.util.RestrictPersonnelNumber;
 import pnet.data.api.util.RestrictPreferredUserId;
+import pnet.data.api.util.RestrictRole;
 import pnet.data.api.util.RestrictSupervisorPersonnelNumber;
 import pnet.data.api.util.RestrictTenant;
 import pnet.data.api.util.RestrictUpdatedAfter;
@@ -49,7 +50,7 @@ public class PersonDataFind extends AbstractFind<PersonItemDTO, PersonDataFind>
     RestrictAdvisorAssignmentCompanyId<PersonDataFind>, RestrictAdvisorAssignmentCompanyNumber<PersonDataFind>,
     RestrictAdvisorAssignmentCompany<PersonDataFind>, RestrictAdvisorAssignmentType<PersonDataFind>,
     RestrictAdvisorAssignmentDivision<PersonDataFind>, RestrictUpdatedAfter<PersonDataFind>,
-    CompanyMergable<PersonDataFind>, Scrollable<PersonDataFind>
+    RestrictDatedBackUntil<PersonDataFind>, CompanyMergable<PersonDataFind>, Scrollable<PersonDataFind>
 {
 
     public PersonDataFind(FindFunction<PersonItemDTO> findFunction, List<Pair<String, Object>> restrictItems)